Imported steamvr, first pass at RB Smooth Locomotion
Apparently steamvr was updated last month, and it now has support for the new Unity XR stuff. Importing steamvr in legacy mode threw up a "105 device not found" error for both controllers, so I'm trying it out in the new Unity XR mode. Likely will need to play with some stuff here, I'm not familiar with the new workflow