Rope projectile now lerps rotation to match the collision normal
Got a semi-jank workaround for maintaining the scale of the projectile after attachment. Solution is to have a separate gameobject for the model, and put the colliders on a separate gameobject. We'll see how that goes for when I have my real models in game, but for now it works